Pass self.dtype to lgs_map_sh in convolution kernels
ConvolutionKernel and GaussianConvolutionKernel called lgs_map_sh
without a dtype, so kernels were always float32 regardless of the
object's precision. The kernel hash already includes the dtype, so
cached kernels stay separate per precision.

Update the memory deallocation test to use the kernel dtype size
instead of assuming float32.

+++ specula/data_objects/convolution_kernel.py
@@ -237,7 +237,8 @@ class ConvolutionKernel(BaseDataObj):
         self.real_kernels = lgs_map_sh(
             self.dimx, self.pupil_size_m, self.launcher_pos, zfocus, layer_offsets,
             self.zprofile, self.spot_size, self.pxscale, self.dimension,
-            overs=self.oversampling, theta=lgs_tt, doCube=True, xp=self.xp
+            overs=self.oversampling, theta=lgs_tt, doCube=True,
+            dtype=self.dtype, xp=self.xp
         )
 
         # Process the kernels - apply FFT if needed
